+++ CHANGES.rst
@@ -2,10 +2,17 @@ Changelog of lizard-measure
 ===================================================
 
 
-1.9 (unreleased)
-----------------
+1.10 (unreleased)
+-----------------
 
 - Require higher lizard-history version.
+
+
+1.9 (2012-02-24)
+----------------
+
+- Adds initial support for suitable measures (beta),
+  issue lizardsystem/lizard-portal#18.
 
 
 1.8 (2012-02-17)